requirements.txt
appdirs==1.4.4
asgiref==3.7.2
astroid==3.0.0
attrs==22.1.0
bottle==0.12.23
build==1.2.1
CacheControl==0.14.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
cleo==2.1.0
crashtest==0.4.1
cryptography==36.0.2
dill==0.3.7
distlib==0.3.8
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
dulwich==0.21.7
erdpy==2.1.0
fastjsonschema==2.19.1
filelock==3.14.0
hidapi==0.12.0.post2
idna==3.4
importlib_metadata==7.1.0
installer==0.7.0
isort==5.12.0
jaraco.classes==3.4.0
keyring==24.3.1
ledgercomm==1.1.2
Markdown==3.5
mccabe==0.7.0
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
pexpect==4.9.0
pkginfo==1.10.0
platformdirs==3.11.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prettytable==3.5.0
ptyprocess==0.7.0
pycparser==2.21
pycryptodomex==3.16.0
pylint==3.0.1
PyNaCl==1.5.0
pyproject_hooks==1.1.0
pytz==2023.3.post1
rapidfuzz==3.9.1
requests==2.28.1
requests-cache==0.9.7
requests-toolbelt==1.0.0
semver==2.13.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
tomlkit==0.12.1
trove-classifiers==2024.5.17
url-normalize==1.4.3
urllib3==1.26.13
virtualenv==20.26.2
wcwidth==0.2.5
xattr==1.1.0
zipp==3.18.2
django-cors-headers==4.4.0
gunicorn==20.0
python-dotenv==1.0.1
whitenoise==6.7.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
dj-database-url==2.2.0
psycopg2==2.9.9
pydash==8.0.4
openapi-spec-validator==0.7.1